Integrated twitter api with PDF converter
The main integration is complete but I will enhance the pdf design, page
layout and such.

I commented out the console.log's in the twitter module,
  fixed a bug in tweet-utils.js > getTweetArray function, it was
  returning undefined instead of an array sometimes
Also in twitter module I changed index.js to main.js, main.js to index.js
because it would be better to export functions in index.js for usage.

I converted all import syntax to require, because currently we're using
plain nodejs, in the future we could all use babel. But for now a
standart should be used to avoid problems.

Added example.env file, it should be replaced to .env for real usage,
      but don't delete example.env it should be always there for new
      cloners of the repo, and never commit .env file (it is already in
          gitignore so please don't change that)